updated modspeak and embed to not do anything if not args are given

+++ commands_to_load/Mod.py
@@ -33,6 +33,9 @@ class Mod():
     @commands.command(aliases=['em'])
     async def embed(self, ctx, *arg):
         logger.info('[Mod embed()] embed function detected by user ' + str(ctx.message.author))
+        if arg == '':
+            logger.error("[Mod embed()] no args, so command ended")
+            return
         await ctx.message.delete()
         logger.info('[Mod embed()] invoking message deleted')
         
@@ -64,6 +67,9 @@ class Mod():
     @commands.command(aliases=['warn'])
     async def modspeak(self, ctx, *arg):
         logger.info('[Mod modspeak()] modspeack function detected by minion ' + str(ctx.message.author))
+        if arg == '':
+            logger.error("[Mod modspeak()] no args, so command ended")
+            return
         await ctx.message.delete()
         logger.info('[Mod embed()] invoking message deleted')
